subProj 1 Virtual CAN Bus Framework/requirements.txt
python-can==4.3.1
pytest==8.0.0
pytest-html==4.1.1
